add the ability to use newer versions of Go (#155)
This provides a means to have go_repositories() load what rules_go needs
to work, but also provide a way for the user to specify newer versions
of Go. rules_go may not be able to keep up with the pace of change that
the user needs in the face of Go bugs or security problems.

This implementation takes a cue from rules_closure by adding a omit_go
boolean to go_repositories.