add cgo support  (#166)
* add cgo support to gazelle

This adds cgo support to gazelle.

A cgo_library rule called cgo_default_library is created if there are
any CFiles, HFiles, or CgoFiles in the package. If there are no GoFiles,
a go_library without srcs is added but with library set to
cgo_default_library.

If a cgo_library was created, we move the assembly SFiles over to the
cgo_library to make sure they are in scope for that code.

This change also makes sure we include the Go-expected capital S
extension of assembly files.

This change depends on the main sandwich PR #164 and the cgo_library in
root of repo workaround #160.