Gazelle select (4/x): convert and maps to select expressions (#451)
newValue now converts maps to to a call to the select function with a
dictionary argument. The keys of the dict are the keys of the map
converted to strings. The values are converted recursively with
newValue.

newValue also supports packages.PlatformStrings. This is converted
into a list if it only has generic strings, a select call if it only
has platform-specific strings, or a concatenation of the two if it has
both.

These cases are not used yet, since rules doesn't use packages.Package
yet. This prepares for later PRs.

Related: #409