Gazelle select (7/x): platform constraints (#494)
Add a table of platforms Gazelle should generate BUILD files for. Each
platform has a label of a predefined config_setting rule and a list of
tags that are true on that platform. 

At this time, platforms are not configurable, but they may be in the
future if we find a compelling use case.

Related #409