Rename import symbol bzl to bf in all tools (#589)
This is in response to a review comment on an earlier Gazelle PR. The
package is "github.com/bazelbuild/buildtools/build", but "build" is
easy to confuse with "go/build". "bzl" is the suffix for Skylark files
(containing rule and macro definitions), so it doesn't seem
appropriate. "bf" can stand for "BUILD file" or "Bazel format".