Gazelle: refactor resolve_external.go to eliminate global state (#590)
The repo root cache and the repoRootForImportPath function pointer are
now fields of externalResolver instead of global. This reduces timing
dependencies between test cases.

Note that currently, only one instance of externalResolver is
created. If we go back to having multiple instances, we'll need to
find a way to share the cache.