Gazelle: flat mode (#723)
When given the -experimental_flat flag on the command line, Gazelle
will generate a single BUILD.bazel file at the top of the repository
that contains rules for everything in the repository.

This flag is intended to be used on external repositories. Files
generated with this flag may be checked in and used with
new_git_repository or new_http_archive. We hope to eventually remove
go_repository.

Related to #460