Gazelle: Fix substitution of ${SRCDIR} in cgo directives (#740)
"go build replaces ${SRCDIR} in cgo directives with the absolute
path of the source directory. Gazelle did the same thing when
extracting copts from source files. This caused builds to be
non-hermetic when the path was used in defines.

With this change, ${SRCDIR} is replaced with the relative path from
the repo root. This will probably break some projects, requiring some
manual tweaking of BUILD files, but hermeticity is more important.

Fixes #704