Compile archives (not objects) and pack extra objects only when needed (#777)
The Go compiler can produce either an .a file (with -pack; contains
export data and compiled code) or an .o file (embeds export data and
compiled code). Additional .o files may be packed into an .a file.

The Go rules have not previously used -pack. We used "go tool pack c"
whether we had extra objects or not to simplify our logic. This is not
safe, since the Go compiler does not embed export data in .o files in
a format that can safely be extracted.

With this change, we always pass -pack to the compiler to produce an
.a file directly. If we have extra objects, we copy it, then append to
the copy. This is what "go build" does (minus the copy).

Fixes #769