Gazelle: delete empty rules (#805)
Generator (now a struct, not an interface) now produces a list of
empty rules. For example, if a go_library was produced but no
go_binary, a go_binary rule will be in the empty list.

MergeWithExisting now accepts the empty list and merges with old
rules. If nothing is left after a rule is merged, it is deleted. Rules
with # keep comments and unmergeable attributes will not be deleted.

Fixes #332