Gazelle: ignore "cgo" tag (#1073)
Previously, rules_go always considered the "cgo" tag to be
true. Gazelle excluded files with !cgo.

Now that we support "pure" builds, the tag may be true or false
depending on the build mode. With this change, Gazelle will ignore the
cgo tag when deciding whether to include a file.

Fixes #1037