Inherit importmap from embedded libraries (#1541)
This change allows Go rules to inherit the importmap attribute from
embedded libraries, the same way the importpath attribute is
inherited. This is particularly important for tests, which rely on an
embedded library under test to set the importpath. Test archives are
imported from the generated testmain, so their importmap does matter.

Fixes #1538