Windows fixes (#1572)
* Use `repository_ctx.template` to copy files.

`cp` is not a recognized command on Windows.

* Fix filter_buildid.go to work on Windows.

`syscall.Exec` on Windows always fails as it's not supported.

* Fix path comparison on Windows.

Bazel often passes paths as Unix paths, even on Windows.

* Use params files for the linker.

If the command line gets too large, because of many deps, bazel will
now pass the arguments to the link.go binary as a params file.

* Use a temp location in pack.go.

Instead of clubbering all over the execroot, extract `.o` files into
a temporary location and clean up afterwards. This also fixes a race
condition if multiple pack operations are done simulataneously, each
containing same-named `.o` files (always the case with cgo).

* Pass the search path in absolute form.

This is to work around the limitation in go on Windows where long
filenames are only supported if they are absolute.