Pure mode when using the goos attribute (#1864)
Currently, the goos attribute on a go_binary rule is ignored when determining whether to force pure mode. The only check being performed right now is using the toolchain target.

This change forces pure mode for a binary when the toolchain target is the same as the host, but the specific binary is cross-compiled using the goos attribute. All other cases remain the same.