go_proto_library: support adjusted import prefixes (#1869)
In Bazel 0.22.0, proto_library will support import_prefix and
strip_import_prefix attributes. This change allows targets with these
attributes to be compiled with go_proto_library.

When invoking protoc, we pass in paths to binary proto descriptors
(the output of proto_library) and a list of .proto names. protoc loads
the descriptors corresponding to the names we pass in. Those names
must match the adjusted import paths for the .proto files.

Fixes #1851