change import_path_option default value to false (#1923)
Fixes #1596

import_path_option is a go_proto_compiler attribute with the following definition:
When true, the importpath attribute from go_proto_library rules using this compiler will be passed to the compiler on the command line as --option import_path={}.

When a proto file does not contain option go_package, go_protobuf and gogoprotobuf use the import_path parameter to determine the pb.go file's package name. When the import_path parameter is not passed in, the proto package name is used to determine the pb.go file's package name.

With the lack of option go_package, it appears to be the expected behaviour to use the proto package name for determining the generated go file's package name. Additionally, the import_path parameter is deprecated, and other go proto compilers do not have an import_path option. This is leading to package name mismatch compiler errors when a go_proto_library rule contains multiple compilers and some are not gogoprotobuf/go_protobuf compilers.

Therefore, I am proposing that the default behavior of the go_proto_compiler rule should be to not pass in the import_path option. Another solution could be to add a new set of rules for gogoprotobuf and go_protobuf that have the import_path_option set to false.