nogo: make fact files optional for dependencies (#1941)
nogo now allows imported archives not to have fact files (.x
files). Fact files are produced for libraries analyzed by nogo. We
already had a special case for standard libraries (which are not
analyzed by nogo), but we also need to support go_tool_library
dependencies, especially //go/tools/coverdata, which is an implicit
dependency when coverage is enabled.

Fixes #1940