go_bazel_test: wrapper and test library for running Bazel from go_test (#2109)
go_bazel_test is a macro that wraps go_test. It will be used for
integration tests in rules_go as a replacement for bazel_test. The
macro adds dependencies on bazel_testing and runfiles from rules_go.

bazel_testing is a new test framework. Its TestMain function may be
called to construct a Bazel workspace (specified as a txtar
archive). Tests are run within this workspace. Bazel may be invoked
with the RunBazel function.

Updates #1508
Updates #1959