GoCompilePkg: include export headers in cgo_exports when needed (#2130)
For packages built with cgo in "c-archive" or "c-shared" modes, we'll
now include the generated "_cgo_export.h" (a.k.a. "_cgo_install.h")
headers in GoArchive.cgo_exports. This lets them be picked up by
go_binary_c_archive_shared, which concatenates all of them with a
genrule. That probably isn't correct (it's not what the Go command
does), but it restores old behavior at least.

Also, fixed tests which were happy with an empty header.

Fixes #2128