go_toolchain: move cgo_context_data dep to go_context_data
Previously, we defined separate cgo and non-cgo variants of
go_toolchain. The cgo variant depended on cgo_context_data, which
requires a C/C++ toolchain. However, go_toolchain is built in host
mode, so we always got a toolchain for the execution platform, not the
target platform. This broke cross-compilation with cgo.

With this change, we only declare a go_toolchain per platform,
regardless of cgo. go_context_data depends on cgo_context_data
if the cgo_off constraint is not set.

This means a C/C++ toolchain (for the correct platform) will be
required for the default target platform and platforms that have not
specifically opted out of cgo. Platforms declared in //go/toolchain
are already fine. However, users defining their own platforms may be
broken by this change.

Fixes #2166