Remove dependency on skylib old_sets.bzl (#2220)
Skylib broke compatibility by renaming this module (from sets.bzl) and
replacing it with something that doesn't handle depsets (which is what
use we use it for). old_sets.bzl is removed to be removed in a future
release.

We don't really need this module anymore. Just use the depset
constructor directly.

Fixes #2157