go/tools/bazel_testing: don't panic on setup error (#2237)
setupWorkspace returned a nil cleanup function if it encountered an
error. It's caller panicked when calling this before checking for an
error. Now, setupWorkspace returns a trivial cleanup function instead.

Also, cleanup now returns an error. If cleanup returns an error, TestMain
prints it, but does not fail (if the test didn't fail).

Updates #2233