tests/integration/popular_repos: disable tests incompatible with 1.15 (#2554)
All of these tests fail to link because the external test package
indirectly imports the internal test package. See #1877 for the
original issue.

This pattern is difficult to implement in Bazel, since it requires an
aspect to recompile test dependencies. It's generally possible to
work around it, so now this fails with a descriptive error message.

Fixes #2535